refactor(core): extract wasm core utility exports
