+++ crates/tsz-core/src/api/wasm/mod.rs
@@ -1,1 +1,2 @@
 pub(crate) mod compiler_options;
+pub(crate) mod core_utils;
